Ignore out-of-sequence chunk checksum error
The process of uploading file chunks for a large file can occasionally
result chunk checksums being stored before the first checksum has been
stored. This results in an error, since the resulting checksum array
can't be parsed by pq.Array without a 0 index available in the array.
This shouldn't actually indicate an error for the upload, however, since
this isn't an error with storing the checksum, and doesn't impact future
chunk uploads.