Filter checksums array from db during upload
Checksums arrays returned by the db can't be parsed by pq.Array if there
are nil values, which occurs when out of sequence chunks are added with
gaps between them (ie chunk 2 and chunk 5 will have chunks 3 and 4 set
to nil in the db). This change removes the nil values before returning,
which should avoid the pq.Array parsing error.