Skip lock creation for B2 re-auth cron task
The Backblaze B2 re-authentication cron task needs to run on every
server, therefore it doesn't need to create an advisory lock in the
database.

This should fix an issue in multi-server setups where one server isn't
able to re-authenticate with Backblaze, leading to an expired token
being used on that server.