[CLI] Remove auto-logout for failed session, add separate debug mode
Automatically removing all session files (effectively logging a user
out) is not the most user friendly way to approach failed session
validation for the CLI, and doesn't really increase security in a
notable way.

This removes the auto-logout and prompts the user to check their
YEETFILE_CLI_KEY env var or log in again.

Debug mode is also now explicitly enabled using debug_mode: true in the
config.yml file. This allows debug messages to be printed outside of the
file, such as printing a limited amount of debug info regarding the CLI
key.

See #33