Implement server invite codes (#37)
This introduces the ability for a YeetFile instance admin to send out
invites to a list of emails with unique signup codes. These signup
codes are tied to the email they're sent to, and are removed after a
user completes registration using that email and signup code.

The invite codes are not stored in plaintext anywhere. They are hashed
with bcrypt before storing in the database, and are only included in
plaintext in the outbound email.

Enabling invite codes requires the following conditions:

- YEETFILE_ALLOW_INVITES set to 1
- YEETFILE_SERVER_PASSWORD set to a non-empty string
- YEETFILE_EMAIL_* variables set